[Minor overhaul] 'if' now evaluates the first element for non-NIL state, and evaluates the first case, otherwise evaluates the second case.   Also, if no second case is supplied, it returns NIL.
Empty SExpr now equates to NIL.

Added the '<=>' operator for three-way comparison and cmp operator to return three-way value (-, 0, +). Ex:
    (<=> (cmp a b) {> "greater"} {< "less"} {"equal"})
The '<=>' operator's first parameter is an expression that equates to a number (can be an SExpr).
  The 'body' expressions are QExprs of two parts, the first part is the comparisons to match (two at most)
    and the second is the expression to evaluate. Ex: {> (+ 5 4)}
  If there are two or more body expressions, the last one can omit the comparison if it is the "otherwise" case,
    meaning that it will match all remaining comparisons.
  If there are 3 body expressions, then the last will always omit the comparison.
  Duplicate comparison values in the body expressions are considered an error so:
    (<=> (cmp a b) {> "gt"} {>= "ge"})
    ...would error
  Any cases not covered evaluate to NIL.

T, NIL, and 'exit' are special-case symbols now, with 'exit' returning a special LVal used to exit the interpreter.

Updated all comparison operators to return T or NIL, which includes all the global logical operators 'or' 'and' 'xor' and 'not'.

Renamed globals.lisp to globals.dl (making .dl the new danlang file extension),
  and modified the 'load' builtin to append the ".dl" and prepend "lib\" if the filename supplied is not found.

Added a number of builtins, including fast-fibonacci calculation.
Also added some parameter checking for builtins (not complete), and renamed most everything to match C# style (generally).
  There is more work to do there, but generally it is a lot better.